refactor: relocate extension execution-context resolution to core
Dissolves the execution-context edge - the second-biggest core->extension
behavior coupling. Like the manifest store (#8852) and scope/invocation-context
(#8854), the ExtensionExecutionContext resolution layer is core glue misfiled in
the extension module: it bridges Component/RunDir (core) with the contract
ExtensionManifest/ExtensionCapability, loading via the now-core manifest store.

Relocates from extension/capability.rs to crate::extension_execution:
ExtensionExecutionContext, resolve_execution_context(_if_available/_for_project),
resolve_extension_for_capability, execution_context_for_extension,
has_linked_extension_for_capability, extract_component_extension_settings,
path_list_env_value, stderr_tail, and the error/guidance helpers - 16 items,
verified ZERO extension-module dependencies.

Left in extension/capability.rs: build_scenario_runner + ScenarioRunnerOptions
(the only ExtensionRunner-touching code) and the one integration test that
exercises ExtensionRunner::run. The resolve-family unit tests moved with their
code. Core re-exports everything so all crate::extension::* paths stay stable.

The 3 core callers (component/resolution, engine/execution_context,
deps_provider) now resolve execution contexts against a core module - no longer
reaching into extension behavior.
